Amplitude modulation (AM) signal processing
This application "AMonUSRP" is based on a file am_usrp710.dat obtained on the Web, with complex-valued samples (the complex envelope of the AM RF signal). I thought it would be easier for the students to first deal with a real-valued signal. Then I created another related application called "AMDemodulation".